fix(imhex): use correct casing for imhex.exe.gui
According to the [chocolatey docs](https://docs.chocolatey.org/en-us/create/create-packages#how-do-i-set-up-shims-for-applications-that-have-a-gui), the gui shim should have the same casing as the executable. The executable's name is imhex.exe.
Also there is the imhex-gui.exe which actually contains the gui functionality.

The imhex.exe is a CLI application that will automatically start the GUI if it is being run from the explorer. This is not possible with chocolatey so even adding this gui shim will unfortunately not help hiding the command line window when calling imhex.exe.